Moving graphql requests from mutations file to requests file
